[✨feat] js 리스트 전체선택 기능 추가 #40
