+++ client/js/kyungseob.js
@@ -56,3 +56,118 @@ export let swiper = new Swiper('.mySwiper', {
   },
   spaceBetween: 10,
 });
+
+export function allSelector(){
+  const totalSelector = document.querySelector('.all-check-box, .all-check-box active')
+  const singleSelector = document.querySelectorAll('.single-check-box', '.single-check-box active')
+
+  totalSelector.addEventListener('click',act)
+
+  function act(){
+    if(totalSelector.className === 'all-check-box active'){
+      for (let index = 0; index < singleSelector.length; index++) {
+        const element = singleSelector[index];
+        element.className = 'single-check-box'
+      }
+      totalSelector.className = 'all-check-box'
+    } else {
+      for (let index = 0; index < singleSelector.length; index++) {
+        const element = singleSelector[index];
+        element.className = 'single-check-box  active'
+      }
+      totalSelector.className = 'all-check-box active'
+      }
+    }
+  }
+
+
+
+
+
+/* export function allCheckBox(){
+  const allSelector = document.querySelector('.all-check-box--true, .all-check-box--false');
+  
+  allSelector.addEventListener('click', changeCheck) 
+
+function changeCheck (){
+  if(allSelector.className === 'all-check-box--false') {
+   return allSelector.className = 'all-check-box--true'
+  } else {
+   return allSelector.className = 'all-check-box--false'
+  }
+ }
+} */
+
+/* export function allCheckBox(){
+  const allSelector = document.querySelector('.all-check-box--true, .all-check-box--false');
+  
+  allSelector.addEventListener('click', changeCheck) 
+
+function changeCheck (){
+  if(allSelector.className === 'all-check-box--false') {
+   return allSelector.className = 'all-check-box--true'
+  } else {
+   return allSelector.className = 'all-check-box--false'
+  }
+ }
+}
+ */
+
+export function coldListAct(){
+  const viewMore = document.querySelector('.vector-1')
+  const coldList = document.querySelector('.cart-list__cold--group')
+  
+  viewMore.addEventListener('click',show)
+
+  function show (){
+    if(viewMore.className === 'rotate-deg--bottom'){
+      return viewMore.className = 'rotate-deg--top', coldList.style.display = 'none';
+    } else {
+      return viewMore.className = 'rotate-deg--bottom', coldList.style.display = 'block';
+    }
+    
+  }
+}
+
+export function frozenListAct(){
+  const viewMore = document.querySelector('.vector-2')
+  const frozenList = document.querySelector('.cart-list__frozen--group')
+  
+  viewMore.addEventListener('click',show)
+
+  function show (){
+    if(viewMore.className === 'rotate-deg--bottom'){
+      return viewMore.className = 'rotate-deg--top', frozenList.style.display = 'none';
+    } else {
+      return viewMore.className = 'rotate-deg--bottom', frozenList.style.display = 'block';
+    }
+    
+  }
+}
+
+export function temperatureListAct(){
+  const viewMore = document.querySelector('.vector-3')
+  const temperatureList = document.querySelector('.cart-list__temperature--group')
+  
+  viewMore.addEventListener('click',show)
+
+  function show (){
+    if(viewMore.className === 'rotate-deg--bottom'){
+      return viewMore.className = 'rotate-deg--top', temperatureList.style.display = 'none';
+    } else {
+      return viewMore.className = 'rotate-deg--bottom', temperatureList.style.display = 'block';
+    }
+    
+  }
+}
+
+export function deleteList(){
+  const deleteIcon = document.querySelector('.group-list__close')
+  const list = document.querySelector('.group-list')
+
+  deleteIcon.addEventListener('click',isDelete)
+
+  function isDelete (){
+    return list.style.display = 'none'
+  }
+}
